Clearer removal of assets
