+++ ui/utils/swap.ts
@@ -1,12 +1,10 @@
 import {
-  AnyAsset,
   isSmartContractFungibleAsset,
   SwappableAsset,
 } from "@tallyho/tally-background/assets"
 import { EIP_1559_COMPLIANT_CHAIN_IDS } from "@tallyho/tally-background/constants"
 import { fixedPointNumberToString } from "@tallyho/tally-background/lib/fixed-point"
 import logger from "@tallyho/tally-background/lib/logger"
-import { normalizeEVMAddress } from "@tallyho/tally-background/lib/utils"
 import { EVMNetwork } from "@tallyho/tally-background/networks"
 import { fetchSwapPrice } from "@tallyho/tally-background/redux-slices/0x-swap"
 import { selectCurrentNetwork } from "@tallyho/tally-background/redux-slices/selectors"
@@ -19,6 +17,7 @@ import {
 import { debounce, DebouncedFunc } from "lodash"
 import { useState, useRef, useCallback } from "react"
 import { CompleteAssetAmount } from "@tallyho/tally-background/redux-slices/accounts"
+import { isSameAsset } from "@tallyho/tally-background/redux-slices/utils/asset-utils"
 import { useBackgroundDispatch, useBackgroundSelector } from "../hooks"
 import { useValueRef, useIsMounted, useSetState } from "../hooks/react-hooks"
 
@@ -126,32 +125,6 @@ export const fetchQuote = async ({
 }
 
 export type QuoteType = "getSellAmount" | "getBuyAmount"
-
-// FIXME Unify once asset similarity code is unified.
-export function isSameAsset(asset1: AnyAsset, asset2: AnyAsset): boolean {
-  if (typeof asset1 === "undefined" || typeof asset2 === "undefined") {
-    return false
-  }
-
-  if (
-    isSmartContractFungibleAsset(asset1) &&
-    isSmartContractFungibleAsset(asset2)
-  ) {
-    return (
-      normalizeEVMAddress(asset1.contractAddress) ===
-      normalizeEVMAddress(asset2.contractAddress)
-    )
-  }
-
-  if (
-    isSmartContractFungibleAsset(asset1) ||
-    isSmartContractFungibleAsset(asset2)
-  ) {
-    return false
-  }
-
-  return asset1.symbol === asset2.symbol
-}
 
 type RequestQuoteUpdateConfig = {
   type: QuoteType
